requirements.txt
scikit-learn 
plotly 
pandas 
matplotlib 
seaborn
numpy
opencv-python
openai==0.28
aiofiles
python-dotenv